+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 **New Features**
 
-- Find tags similar to given value
+- Find tags with `tag` value similar to given keyword
+- Find recipient with `name` or `metadata` containing given keyword
 
 ## v2.3.0 [October 23, 2017]